pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "confit"
version = "0.1.0"
dependencies = [
    "pyyaml>=6.0.2",
]

[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
test = [
    "pytest>=8.3.4",
]